Deleted my first test and created a new one in my group's class
